Initial Commit
This commit creates a new repository with the API Payload schemas for
interacting with the IBMQ devices. These schemas are the source of truth
for validating these payloads. The schema files in this repo are copied
from the current state of the qiskit-terra repo and moving forward these
should be used anywhere that the copies in terra were previously being
used.