Attempt to fix ci job triggers (#2)
* Attempt to fix ci job triggers

The repo was created with a main branch instead of master. However since
the CI was copied from an older repo using a master branch the triggers
weren't correcly set to use main. This commit attempts to fix this by
updating the trigger branches to be main.

* Add missing discover argument to unittest job

Now that CI is running this commit corrects an oversight in the unittest
command. It was missing the discover argument so the job was not running
unittests correctly. This commit corrects the oversight.